requirements.txt
flet==0.8.0
flet_core==0.8.0
dbus_python==1.3.2
PyGObject==3.44.1
